[#818] Add `UserRolePayload` to create and retrieve `UserRole` functions
* Check whether roles are assigned to a user before deleting

+++ GAE/src/org/waterforpeople/mapping/app/web/rest/UserRestService.java
@@ -293,12 +293,12 @@ public class UserRestService {
     @ResponseBody
     public Map<String, Object> createUserRole(@RequestBody UserRolePayload payload) {
         final RestStatusDto statusDto = new RestStatusDto();
-        statusDto.setStatus("failed");
 
         final Map<String, Object> response = new HashMap<String, Object>();
         response.put("meta", statusDto);
 
         if (StringUtils.isBlank(payload.getName())) {
+            statusDto.setStatus("failed");
             statusDto.setMessage("_missing_role_name");
             return response;
         }
@@ -324,7 +324,11 @@ public class UserRestService {
     @ResponseBody
     public Map<String, Object> listUserRoles() {
         final Map<String, Object> response = new HashMap<String, Object>();
-        response.put("roles", userRoleDao.listAllRoles());
+        List<UserRolePayload> rolesPayload = new ArrayList<UserRolePayload>();
+        for (UserRole role : userRoleDao.listAllRoles()) {
+            rolesPayload.add(new UserRolePayload(role));
+        }
+        response.put("roles", rolesPayload);
         return response;
     }
 
@@ -339,14 +343,15 @@ public class UserRestService {
     public Map<String, Object> findUserRole(@PathVariable Long roleId) {
         final Map<String, Object> response = new HashMap<String, Object>();
         RestStatusDto statusDto = new RestStatusDto();
-        statusDto.setStatus("ok");
+        response.put("meta", statusDto);
 
         UserRole role = userRoleDao.getByKey(roleId);
         if (role == null) {
             statusDto.setMessage("_role_not_found");
-            response.put("meta", statusDto);
+            return response;
         }
-        response.put("role", role);
+        statusDto.setStatus("ok");
+        response.put("role", new UserRolePayload(role));
         return response;
     }
 
@@ -410,10 +415,18 @@ public class UserRestService {
         response.put("meta", statusDto);
 
         UserRole deleteRole = userRoleDao.getByKey(roleId);
+        if (deleteRole == null) {
+            statusDto.setStatus("ok");
+            statusDto.setMessage("_role_not_found");
+            return response;
+        }
+
         if (userDao.listUsersByRole(deleteRole.getName()).isEmpty()) {
             userRoleDao.delete(deleteRole);
             statusDto.setStatus("ok");
             statusDto.setMessage("_role_deleted");
+        } else {
+            statusDto.setMessage("_role_in_use");
         }
 
         return response;
